pyproject.toml
[project]
name = "octopus-bot"
version = "0.1.0"
description = "Telegram bot for Octopus servers management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
    "psutil>=6.0",
    "schedule>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "coverage>=7.0",
]
